Time analysis
Rearranged the notebook so that we merge the training dataset with the riders dataset. It seemed like there was a lot of repetition when doing the EDA and it didn't make sense to look at the two datasets seperatley if we're going to merge them in the end. I also changed the order of the columns so that the column with the response variable is at the end of the dataframe.
I created a new dataframe with the hour of order placement, confirmation and  pickup to see if there are any delays that might affect the final delivery time.